requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0
sqlalchemy>=2.0.0